+++ src/parsers/problem/PTAProblemParser.ts
@@ -16,9 +16,9 @@ export class PTAProblemParser extends Parser {
 
     task.setName(container.querySelector('span.font-bold').textContent.trim());
 
-    const limits = container.querySelectorAll('div[class*="problemInfo"] .pc-text-raw');
-    const timeLimitStr = limits[3].textContent;
-    const memoryLimitStr = limits[5].textContent;
+    const limits = [...container.querySelectorAll('div[class*="problemInfo"] .pc-text-raw')].map(l => l.textContent);
+    const timeLimitStr = limits.find(text => text.includes('MB'));
+    const memoryLimitStr = limits.find(text => text.includes('ms'));
 
     task.setTimeLimit(parseInt(/(\d+)/.exec(timeLimitStr)[1], 10));
     task.setMemoryLimit(parseInt(/(\d+)/.exec(memoryLimitStr)[1], 10));