+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-## Unreleased
+## 2.43.0 (Unreleased)
+- Added support for AlgoZenith
 - Fixed the parsing of TLX course problems and problems with time limits in milliseconds (thanks [@ChrisG661](https://github.com/ChrisG661))
 - Fixed the parsing of Codeforces problems with time/memory limits in bold text (thanks [@colaczy](https://github.com/colazcy))
 - Fixed the parsing of PTA problems